third_party/WebKit/LayoutTests/editing/execCommand/non-html-document.html

<!DOCTYPE html>
<script src='../../resources/js-test.js'></script>
<script>
description("Non HTMLDocument such as XMLDocument shouldn't support execCommand");

var xmldoc = window.document.implementation.createDocument('http://www.w3.org/1999/xlink', 'html', null);
shouldThrow("xmldoc.execCommand('bold')");
shouldThrow("xmldoc.queryCommandEnabled('bold')");
shouldThrow("xmldoc.queryCommandIndeterm('bold')");
shouldThrow("xmldoc.queryCommandState('bold')");
shouldThrow("xmldoc.queryCommandSupported('bold')");
shouldThrow("xmldoc.queryCommandValue('bold')");
</script>
