third_party/WebKit/LayoutTests/editing/execCommand/non-html-document-expected.txt

Non HTMLDocument such as XMLDocument shouldn't support execCommand

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xmldoc.execCommand('bold') threw exception InvalidStateError: Failed to execute 'execCommand' on 'Document': execCommand is only supported on HTML documents..
PASS xmldoc.queryCommandEnabled('bold') threw exception InvalidStateError: Failed to execute 'queryCommandEnabled' on 'Document': queryCommandEnabled is only supported on HTML documents..
PASS xmldoc.queryCommandIndeterm('bold') threw exception InvalidStateError: Failed to execute 'queryCommandIndeterm' on 'Document': queryCommandIndeterm is only supported on HTML documents..
PASS xmldoc.queryCommandState('bold') threw exception InvalidStateError: Failed to execute 'queryCommandState' on 'Document': queryCommandState is only supported on HTML documents..
PASS xmldoc.queryCommandSupported('bold') threw exception InvalidStateError: Failed to execute 'queryCommandSupported' on 'Document': queryCommandSupported is only supported on HTML documents..
PASS xmldoc.queryCommandValue('bold') threw exception InvalidStateError: Failed to execute 'queryCommandValue' on 'Document': queryCommandValue is only supported on HTML documents..
PASS successfullyParsed is true

TEST COMPLETE
